fix: move save_image function to new robotoff.images module
